some cleanup work in cart models
