+++ cart/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from django.contrib.admin import ModelAdmin
-from .models import Cart, Orders
+from .models import Cart, Order
 
 
 @admin.register(Cart)
@@ -8,6 +8,6 @@ class CartAdmin(ModelAdmin):
     pass
 
 
-@admin.register(Orders)
+@admin.register(Order)
 class OrdersAdmin(ModelAdmin):
     pass
